+++ src/components/Form.js
@@ -1,15 +1,12 @@
 import React, {Component} from 'react';
 import PropTypes from 'prop-types';
+import EventEmitter from 'eventemitter2';
 import AllComponents from 'formiojs/components';
 import Components from 'formiojs/components/Components';
 Components.setComponents(AllComponents);
-import Form from 'formiojs/Form';
+import FormioForm from 'formiojs/Form';
 
-export default class extends Component {
-  static defaultProps = {
-    options: {}
-  };
-
+export default class Form extends Component {
   static propTypes = {
     src: PropTypes.string,
     url: PropTypes.string,
@@ -20,7 +17,6 @@ export default class extends Component {
       noAlerts: PropTypes.boolean,
       i18n: PropTypes.object,
       template: PropTypes.string,
-      templates: PropTypes.any,
     }),
     onPrevPage: PropTypes.func,
     onNextPage: PropTypes.func,
@@ -30,25 +26,39 @@ export default class extends Component {
     onSubmitDone: PropTypes.func,
     onFormLoad: PropTypes.func,
     onError: PropTypes.func,
-    onRender: PropTypes.func
+    onRender: PropTypes.func,
+    formioform: PropTypes.any
   };
 
+  static getDefaultEmitter() {
+    return new EventEmitter({
+      wildcard: false,
+      maxListeners: 0
+    });
+  }
+
   componentDidMount = () => {
-    const {options, src, url, form} = this.props;
+    const {options = {}, src, url, form} = this.props;
+
+    if (!options.events) {
+      options.events = Form.getDefaultEmitter();
+    }
 
     if (src) {
-      this.createPromise = new Form(this.element, src, options).then(formio => {
+      this.createPromise = new (this.props.formioform || FormioForm)(this.element, src, options).ready.then(formio => {
         this.formio = formio;
         this.formio.src = src;
       });
     }
     if (form) {
-      this.createPromise = new Form(this.element, form, options).then(formio => {
+      this.createPromise = new (this.props.formioform || FormioForm)(this.element, form, options).ready.then(formio => {
         this.formio = formio;
         this.formio.form = form;
         if (url) {
           this.formio.url = url;
         }
+
+        return this.formio;
       });
     }
 
@@ -82,17 +92,21 @@ export default class extends Component {
   };
 
   componentWillReceiveProps = (nextProps) => {
-    const {options, src, form, submission} = this.props;
+    const {options = {}, src, form, submission} = this.props;
+
+    if (!options.events) {
+      options.events = Form.getDefaultEmitter();
+    }
 
     if (src !== nextProps.src) {
-      this.createPromise = new Form(this.element, nextProps.src, options).then(formio => {
+      this.createPromise = new (this.props.formioform || FormioForm)(this.element, nextProps.src, options).ready.then(formio => {
         this.formio = formio;
         this.formio.src = nextProps.src;
       });
       this.initializeFormio();
     }
     if (form !== nextProps.form) {
-      this.createPromise = new Form(this.element, nextProps.form, options).then(formio => {
+      this.createPromise = new (this.props.formioform || FormioForm)(this.element, nextProps.form, options).ready.then(formio => {
         this.formio = formio;
         this.formio.form = form;
       });