+++ src/index.js
@@ -1,3 +1,3 @@
-export {default as Form} from './components/Form';
-export {default as FormBuilder} from './components/FormBuilder';
-export {default as Grid} from './components/Grid';
+export * from './components';
+export * from './modules';
+export {Components, Formio} from 'formiojs';