+++ src/modules/form/constants.js
@@ -1,0 +1,5 @@
+export const FORM_REQUEST = 'FORM_REQUEST';
+export const FORM_SUCCESS = 'FORM_SUCCESS';
+export const FORM_FAILURE = 'FORM_FAILURE';
+export const FORM_SAVE = 'FORM_SAVE';
+export const FORM_RESET = 'FORM_RESET';