+++ src/modules/form/selectors.js
@@ -1,0 +1,3 @@
+import {selectRoot} from '../root/selectors';
+
+export const selectForm = (name, state) => selectRoot(name, state).form;