+++ src/modules/root/selectors.js
@@ -1,0 +1,1 @@
+export const selectRoot = (name, state) => state[name];