+++ src/modules/submission/selectors.js
@@ -1,0 +1,3 @@
+import {selectRoot} from '../root/selectors';
+
+export const selectSubmission = (name, state) => selectRoot(name, state).submission;