+++ src/modules/submissions/selectors.js
@@ -1,0 +1,3 @@
+import {selectRoot} from '../root/selectors';
+
+export const selectSubmissions = (name, state) => selectRoot(name, state).submissions;